Creates controllers for routes
